+++ lib/workato/connector/sdk/dsl/workato_package.rb
@@ -2,11 +2,9 @@
 # frozen_string_literal: true
 
 require 'jwt'
-require_relative './csv_package'
-require_relative './net_package'
-require_relative './stream_package'
-
-using Workato::Extension::HashWithIndifferentAccess
+require_relative 'csv_package'
+require_relative 'net_package'
+require_relative 'stream_package'
 
 module Workato
   module Connector
@@ -72,9 +70,9 @@ module Workato
               end
             end
 
-            header_fields = HashWithIndifferentAccess.wrap(header_fields)
-                                                     .except(:typ, :alg)
-                                                     .reverse_merge(typ: 'JWT', alg: algorithm)
+            header_fields = Utilities::HashWithIndifferentAccess.wrap(header_fields)
+                                                                .except(:typ, :alg)
+                                                                .reverse_merge(typ: 'JWT', alg: algorithm)
 
             ::JWT.encode(payload, key, algorithm, header_fields)
           rescue JWT::IncorrectAlgorithm
@@ -104,6 +102,8 @@ module Workato
             raise Sdk::ArgumentError, 'Mismatched algorithm and key'
           rescue OpenSSL::PKey::PKeyError
             raise Sdk::ArgumentError, 'Invalid key'
+          rescue JWT::VerificationError
+            raise Sdk::ArgumentError, 'Invalid signature'
           end
 
           def verify_rsa(payload, certificate, signature, algorithm = 'SHA256')
@@ -174,6 +174,8 @@ module Workato
             cipher.key = key
             cipher.iv = init_vector if init_vector.present?
             Types::Binary.new(cipher.update(string) + cipher.final)
+          rescue OpenSSL::Cipher::CipherError => e
+            raise Sdk::ArgumentError, e.message
           end
 
           def pbkdf2_hmac_sha1(string, salt, iterations = 1000, key_len = 16)